# TS: Extract contextual type for object/array literals