--- a/javascript/ql/test/library-tests/TypeScript/NestedLiteral/test.ql
+++ b/javascript/ql/test/library-tests/TypeScript/NestedLiteral/test.ql
@@ -1,0 +1,4 @@
+import javascript
+
+from Expr e
+select e, e.getType()