--- a/javascript/ql/test/library-tests/TypeScript/NestedLiteral/tst.ts
+++ b/javascript/ql/test/library-tests/TypeScript/NestedLiteral/tst.ts
@@ -1,0 +1,30 @@
+interface T {
+  name: string;
+  children?: T[];
+}
+
+let context: T = {
+  name: 'x',
+  children: [
+    { name: 'x1' },
+    {
+      name: 'x2',
+      children: [
+        { name: 'x3' }
+      ]
+    }
+  ]
+}
+
+let nocontext = {
+  name: 'x',
+  children: [
+    { name: 'x1' },
+    {
+      name: 'x2',
+      children: [
+        { name: 'x3' }
+      ]
+    }
+  ]
+}