--- a/javascript/ql/test/library-tests/TypeScript/RegressionTests/EmptyName/test.expected
+++ b/javascript/ql/test/library-tests/TypeScript/RegressionTests/EmptyName/test.expected
@@ -1,4 +1,3 @@
-|  in unknown scope |
 | Array in global scope |
 | Intl in global scope |
 | Intl.CollatorOptions in global scope |