… javascript/ql/test/library-tests/TypeScript/NestedLiteral/tsconfig.json