When an object/array literal has a contextual type, we now extract that instead of its inferred expression type. This was motivated by a performance gains in cherow shown below, but I believe it's also an improvement on its own.

|        | before | after |
|--------|--------|-------|
| cherow | 217s   | 109s  |
| vscode | 288s   | 270s  |

For example:
```typescript
interface TNode {
  x?: number;
  child?: TNode;
}
let obj: TNode = { x: 123 }
```
The object literal above now has type `TNode` instead of the anonymous type `{ x: 123 }`. This means it would be found if searching for all expressions of type `TNode`, which seems useful. Being able to search for expressions with a literal type doesn't seem as important, since the DataFlow IR can handle that anyway.

The main reason for this switch is to speed up extraction of large nested literals like in [this file](https://github.com/cherow/cherow/blob/b0ff0f6386bca6cfbf7231ec9ebbe1d71d8b08c0/test/parser/expressions/class.ts). The type of a large object/array literal is essentially a copy of that object, which led to an N^2 cost in the depth of the object. The contextual type tends to be much smaller, usually just a named type, possibly with some type arguments.

For example, when processing this line
```typescript
let deep: TNode = { child: { child: { child: { x: 123 }}}}
```
Each object literal now has the type `TNode`, previously we would independently process these four types:
```typescript
{ x: 123 }
{ child: { x: 123 }}
{ child: { child: { x: 123 }}}
{ child: { child: { child: { x: 123 }}}}
```
I tried a few ideas to reuse intermediate results when processing these types, but it didn't make much of a difference, and switching to contextual types was way simpler.

Running a sanity check evaluation on typescript-full.slugs.