@esben-semmle 2019-04-08
LGTM.
Would it make sense to investigate changing the inference for functions as well? I.e. could we gain anything by using the declared type of `() => 42` instead of the inferred expression type?