@asger-semmle 2019-04-08
[Sanity check evaluation](https://git.semmle.com/asger/dist-compare-reports/tree/profiling-js-asger.northeurope.cloudapp.azure.com_1554727404365) came out fine.

> Would it make sense to investigate changing the inference for functions as well?

Worth trying at least. I'll try it out this week and see what happens.