@asger-semmle 2019-04-09
> Hmm. This technically requires an extractor bump.

If I understand correctly, it's only used for trap caching, which is disabled for TypeScript.