@esben-semmle 2019-04-09
Right. This is not the first time I have forgotten about that disabled caching. Everything is fine then.