@asger-semmle 2019-04-09
Didn't seem to make a difference for function types.